Fixed Heuristic movement and Q ability aim, added cached inputs for heuristics.
Playing against the best self play agent of 1358 notes: agent is somewhat hard to hit as it moves semi randomly. The agent doesn't always aim towards the player, it manages to hit the player at close distances. If the player stands still it doesn't immediately aim and shoot.
Todo: add curiosity back to the config.
Add walls to the map.
Record demonstration for GAIL