Self-play alone + GAIL against self play agent. Best model yet, it has learned to effectively play the bullet arena.
Both are reinforcement learning agents trained with GAIL (imitation learning) and self play.
bullet_arena_gail_05:
The are correctly killing each other and then going to the goal, but when they're already close to the goal they realise that killing the other is not worth it and go straight towards the goal.